added genres to separate widget
